feat(desktop): add in-app server switching and background updates
- Add servers bridge for connect/list/forget from the server web app
- Use frameless windows with app-region dragging and overlay scrollbars
- Poll releases in background and expose last check time
- Harden capture library manifest, game confidence, and missing-file recovery
- Serve styled login callback page and default bare loopback to HTTP

+++ packages/capture-library/src/types.rs
@@ -40,7 +40,8 @@ pub struct GameGuess {
     pub window_title: Option<String>,
     pub window_class: Option<String>,
     pub icon_url: Option<String>,
-    pub confidence: f64,
+    /// Detector confidence in percent (0-100), matching the recorder protocol.
+    pub confidence: u8,
     pub match_kind: String,
 }
